move label tests closer to implementation
