+++ po/POTFILES.in
@@ -38,6 +38,7 @@ subiquity/common/filesystem/labels.py
 subiquity/common/filesystem/manipulator.py
 subiquity/common/filesystem/tests/__init__.py
 subiquity/common/filesystem/tests/test_actions.py
+subiquity/common/filesystem/tests/test_labels.py
 subiquity/common/filesystem/tests/test_manipulator.py
 subiquity/common/__init__.py
 subiquity/common/serialize.py
